Move basemap and project component source and layer components out of mapUtils file
