+++ moped-editor/src/views/projects/projectView/ProjectComponents/mapUtils.js
@@ -1,7 +1,5 @@
 import { useMemo } from "react";
-import { Source, Layer } from "react-map-gl";
-import { basemaps, SOURCES } from "./mapSettings";
-import { MAP_STYLES, COLORS } from "./mapStyleSettings";
+import { MAP_STYLES } from "./mapStyleSettings";
 
 const mapStyles = MAP_STYLES;
 
@@ -24,269 +22,3 @@ export const interactiveLayerIds = Object.entries(mapStyles).reduce(
   },
   []
 );
-
-/**
- * Component that renders Mapbox source and layers needed for the aerial basemap
- * @param {string} basemapKey
- * @returns JSX.Element
- */
-export const BaseMapSourceAndLayers = ({ basemapKey }) => {
-  return (
-    <>
-      <Source {...basemaps.aerial.sources.aerials} />
-      <Layer
-        {...basemaps.aerial.layers.aerials}
-        layout={{ visibility: basemapKey === "aerial" ? "visible" : "none" }}
-      />
-      {/* Always show street labels so they can be the "target" of beforeId 
-      and always appear on top of everything else */}
-      <Layer
-        {...{
-          ...basemaps.aerial.layers.streetLabels,
-          layout: {
-            ...basemaps.aerial.layers.streetLabels.layout,
-          },
-          // Update street label text color to be readable on either basemap type
-          paint:
-            basemapKey === "aerial"
-              ? {
-                  "text-color": COLORS.white,
-                  "text-halo-color": COLORS.black,
-                  "text-halo-width": 1,
-                }
-              : {
-                  "text-color": COLORS.black,
-                },
-        }}
-      />
-    </>
-  );
-};
-
-/**
- * Component that renders all sources and layers for project components
- * All layers are set to show below basemap street labels using beforeId = "street-labels"
- * @param {Object} data - GeoJSON data for all project components
- * @param {Boolean} isEditingComponent - are we editing a component?
- * @param {String} linkMode - Tracks if we are editing "lines" or "points"
- * @param {Object} clickedComponent - Details of the component that was clicked
- * @param {Object} componentFeatureCollection - GeoJSON data for the component clicked
- * @returns JSX.Element
- */
-export const ProjectComponentsSourcesAndLayers = ({
-  data,
-  isEditingComponent,
-  linkMode,
-  clickedComponent,
-  componentFeatureCollection,
-  isDrawing,
-}) => {
-  // This is a temporary to get data into the map sources
-  const {
-    ctnLinesGeojson,
-    ctnPointsGeojson,
-    projectLines,
-    projectPoints,
-    draftComponentFeatures,
-  } = data;
-
-  const isViewingComponents = !isEditingComponent && !clickedComponent;
-  const isEditingLines =
-    isEditingComponent && linkMode === "lines" && !isDrawing;
-  const isEditingPoints =
-    isEditingComponent && linkMode === "points" && !isDrawing;
-  const shouldShowMutedFeatures = clickedComponent || isEditingComponent;
-  console.log(clickedComponent);
-
-  return (
-    <>
-      <Source
-        id="ctn-lines"
-        type="geojson"
-        data={ctnLinesGeojson}
-        promoteId={SOURCES["ctn-lines"]._featureIdProp}
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["ctn-lines-underlay"].layerProps,
-            layout: {
-              ...mapStyles["ctn-lines-underlay"].layerProps.layout,
-              visibility: isEditingLines ? "visible" : "none",
-            },
-          }}
-        />
-
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["ctn-lines"].layerProps,
-            layout: {
-              ...mapStyles["ctn-lines"].layerProps.layout,
-              visibility: isEditingLines ? "visible" : "none",
-            },
-          }}
-        />
-      </Source>
-
-      <Source
-        id="project-lines"
-        type="geojson"
-        data={projectLines}
-        promoteId="id"
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["project-lines-underlay"].layerProps,
-            layout: {
-              ...mapStyles["project-lines-underlay"].layerProps.layout,
-              visibility: isEditingLines ? "visible" : "none",
-            },
-          }}
-        />
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["project-lines"].layerProps,
-            layout: {
-              ...mapStyles["project-lines"].layerProps.layout,
-              visibility: isViewingComponents ? "visible" : "none",
-            },
-          }}
-        />
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["project-lines-muted"].layerProps,
-            layout: {
-              ...mapStyles["project-lines-muted"].layerProps.layout,
-              visibility: shouldShowMutedFeatures ? "visible" : "none",
-            },
-          }}
-        />
-      </Source>
-
-      <Source
-        id="ctn-points"
-        type="geojson"
-        data={ctnPointsGeojson}
-        promoteId={SOURCES["ctn-points"]._featureIdProp}
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["ctn-points-underlay"].layerProps,
-            layout: {
-              visibility: isEditingPoints ? "visible" : "none",
-            },
-          }}
-        />
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["ctn-points"].layerProps,
-            layout: {
-              visibility: isEditingPoints ? "visible" : "none",
-            },
-          }}
-        />
-      </Source>
-
-      <Source
-        id="project-points"
-        type="geojson"
-        data={projectPoints}
-        promoteId="id"
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["project-points"].layerProps,
-            layout: {
-              ...mapStyles["project-points"].layerProps.layout,
-              visibility: isViewingComponents ? "visible" : "none",
-            },
-          }}
-        />
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["project-points-muted"].layerProps,
-            layout: {
-              visibility: shouldShowMutedFeatures ? "visible" : "none",
-            },
-          }}
-        />
-      </Source>
-
-      <Source
-        id="draft-component-lines"
-        type="geojson"
-        data={draftComponentFeatures}
-        promoteId="id"
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["draft-component-lines"].layerProps,
-            layout: {
-              ...mapStyles["clicked-component-features-lines"].layerProps
-                .layout,
-              visibility: linkMode === "lines" ? "visible" : "none",
-            },
-          }}
-        />
-      </Source>
-
-      <Source
-        id="draft-component-points"
-        type="geojson"
-        data={draftComponentFeatures}
-        promoteId="id"
-      >
-        <Layer
-          beforeId="street-labels"
-          {...mapStyles["draft-component-points"].layerProps}
-          layout={{ visibility: linkMode === "points" ? "visible" : "none" }}
-        />
-      </Source>
-
-      <Source
-        id="clicked-component-features"
-        type="geojson"
-        data={componentFeatureCollection}
-        promoteId="id"
-      >
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["clicked-component-features-lines"].layerProps,
-            layout: {
-              ...mapStyles["clicked-component-features-lines"].layerProps
-                .layout,
-              visibility:
-                componentFeatureCollection &&
-                clickedComponent.line_representation
-                  ? "visible"
-                  : "none",
-            },
-          }}
-        />
-        <Layer
-          beforeId="street-labels"
-          {...{
-            ...mapStyles["clicked-component-features-points"].layerProps,
-            layout: {
-              visibility:
-                componentFeatureCollection &&
-                !clickedComponent.line_representation
-                  ? "visible"
-                  : "none",
-            },
-          }}
-        />
-      </Source>
-    </>
-  );
-};
